+++ src/interfaces/parameters/IHaveParams.php
@@ -1,0 +1,10 @@
+<?php
+namespace extas\interfaces\parameters;
+
+interface IHaveParams
+{
+    public const FIELD__PARAMS = 'params';
+
+    public function getParams(): array;
+    public function buildParams(): IParams;
+}